stub out bridge
This allows using bridge-stub with pre-determined proofs (independant of
L1 block number / time / etc)

+++ rollup/contracts/bridge/Messages.sol
@@ -19,7 +19,7 @@ library Messages {
             keccak256(
                 abi.encodePacked(
                     kind,
-                    sender,
+                    uint256(uint160(sender)),
                     blockNumber,
                     timestamp,
                     inboxSeqNum,
